Fix limit issue where any string showed all tasks
Issue was that there was no check for whether the 'limit:' command was
receiving an integer, and when it wasn't, and the received value wasn't
'all' or 'fit', it defaulted to 'all' based on how task warrior
interpreted 'limit:*' where '*' was any string. So, now it accomodates
for variable inputs and only accepts change if 'limit:' receives an
integer or one of the two keywords.